requirements.txt
streamlit
docs
python-docx